Organize tensors -> Decorators and Interpreters
