+++ syft/frameworks/torch/tensors/__init__.py
@@ -1,6 +1,5 @@
-from .logging import LoggingTensor  # noqa: F401
-from .pointer import PointerTensor  # noqa: F401
-from .abstract import AbstractTensor  # noqa: F401
-from .native import TorchTensor  # noqa: F401
-from .precision import FixedPrecisionTensor  # noqa: F401
-from .add import AdditiveSharingTensor  # noqa: F401
+# Do not change the import sequence it can cause deadlock
+from . import interpreters
+from . import decorators
+
+__all__ = ["decorators","interpreters"]
