+++ syft/frameworks/torch/tensors/decorators/__init__.py
@@ -1,0 +1,1 @@
+from .logging import LoggingTensor  # noqa: F401
