+++ syft/frameworks/torch/tensors/interpreters/__init__.py
@@ -1,0 +1,5 @@
+from .pointer import PointerTensor  # noqa: F401
+from .abstract import AbstractTensor  # noqa: F401
+from .native import TorchTensor  # noqa: F401
+from .precision import FixedPrecisionTensor  # noqa: F401
+from .add import AdditiveSharingTensor  # noqa: F401
