+++ syft/frameworks/torch/tensors/interpreters/add.py
@@ -1,6 +1,6 @@
 import torch
-from syft.frameworks.torch.tensors.abstract import AbstractTensor
-from .utils import hook
+from syft.frameworks.torch.tensors.interpreters.abstract import AbstractTensor
+from syft.frameworks.torch.tensors.interpreters.utils import hook
 
 
 class AdditiveSharingTensor(AbstractTensor):
