+++ test/torch/tensors/test_native.py
@@ -2,7 +2,7 @@ import torch
 import syft
 import random
 
-from syft.frameworks.torch.tensors import PointerTensor
+from syft.frameworks.torch.tensors.interpreters import PointerTensor
 
 
 def test___str__(workers):
