+++ test/torch/test_hook.py
@@ -8,7 +8,7 @@ import random
 
 import syft
 from syft.exceptions import RemoteTensorFoundError
-from syft.frameworks.torch.tensors import PointerTensor
+from syft.frameworks.torch.tensors.interpreters import PointerTensor
 
 
 def test___init__(hook):
